# example/simple-rest-server/scripts/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Simple REST API",
    "version": "1.0.0",
    "description": "一个标准的RESTful演示服务，包含Authors、Books、Notes等资源",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/authors": {
      "get": {
        "summary": "获取所有作者",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "description": "按作者姓名过滤"
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "description": "按邮箱过滤"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "返回结果数量限制"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "分页偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回作者列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Author"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "总数量"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "限制数量"
                    },
                    "offset": {
                      "type": "integer",
                      "description": "偏移量"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建新作者",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "作者姓名"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "作者邮箱"
                  },
                  "bio": {
                    "type": "string",
                    "description": "作者简介"
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date",
                    "description": "出生日期"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功创建作者",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "根据ID获取作者详情",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "作者ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回作者详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "更新作者信息",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "作者ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "作者姓名"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "作者邮箱"
                  },
                  "bio": {
                    "type": "string",
                    "description": "作者简介"
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date",
                    "description": "出生日期"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功更新作者",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "删除作者",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "作者ID"
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除作者"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/authors/{id}/books": {
      "get": {
        "summary": "获取作者的所有书籍",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "作者ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回作者的书籍列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
...
}